labprintconfig: Add command line options and fix group checking
Add command line options to exclude the adding either the BW or the
Color printer, ignoring whether or not the user is currently on the
network, specifying the names of the printers on the server for BW
and for color, and choosing to not configure duplex printing if desired.
Additionally, display help text if an invalid option is provided.

To check which groups a user must be a part of, check the CUPS SystemGroup
directive and see if the user has a group that matches one in that list.